init: move util related functions to their own file
